+++ swagger_server/uas_lib/uas_mgr.py
@@ -25,15 +25,17 @@ Class that implements UAS functions not requiring user attributes
 #pylint: disable=too-many-lines
 
 import json
+import re
 from flask import abort
 from kubernetes import client
 from swagger_server.uas_lib.uas_base import UasBase
-from swagger_server.uas_lib.uas_base import UAIInstance
+from swagger_server.uas_lib.uai_instance import UAIInstance
 from swagger_server.uas_data_model.uai_image import UAIImage
 from swagger_server.uas_data_model.uai_volume import UAIVolume
 from swagger_server.uas_data_model.uai_resource import UAIResource
 from swagger_server.uas_data_model.uai_class import UAIClass
 from swagger_server.uas_data_model.populated_config import PopulatedConfig
+from swagger_server.uas_lib.uas_logging import logger
 
 # pylint: disable=too-many-public-methods
 class UasManager(UasBase):
@@ -48,6 +50,10 @@ class UasManager(UasBase):
         """Delete a list of UAIs optionally selected by class and owner
 
         """
+        logger.debug(
+            "delete UAIs class_id = %s, owner = %s, uai_list = %s",
+            class_id, owner, uai_list
+        )
         self.uas_cfg.get_config()
         labels = []
         if not uai_list:
@@ -55,13 +61,14 @@ class UasManager(UasBase):
                 labels.append("user=%s" % owner)
             if class_id is not None:
                 labels.append("uas-class-id=%s" % class_id)
-            uai_list = self.select_deployments(labels=labels)
+            uai_list = self.select_jobs(labels=labels)
         else:
             uai_list = [
                 uai_name.strip() for uai_name in uai_list
                 if uai_name.strip() != ""
             ]
         resp_list = self.remove_uais(uai_list)
+        logger.debug("uai's deleted: %s'", resp_list)
         return resp_list
 
     def create_uai(self,
@@ -72,6 +79,11 @@ class UasManager(UasBase):
         """Create a new UAI
 
         """
+        logger.debug(
+            "create UAI class_id = %s, owner = %s, passwd_str = %s, "
+            "public_key_str = %s",
+            class_id, owner, passwd_str, public_key_str
+        )
         self.uas_cfg.get_config()
         missing = ""
         if class_id is not None:
@@ -87,16 +99,19 @@ class UasManager(UasBase):
             passwd_str=passwd_str,
             public_key=public_key_str
         )
-        return self.deploy_uai(uai_class, uai_instance, self.uas_cfg)
+        ret = self.deploy_uai(uai_class, uai_instance, self.uas_cfg)
+        logger.debug("uai's created: %s'", ret)
+        return ret
 
     def get_uai(self, uai_name):
         """Retrieve the named UAI
 
         """
+        logger.debug("getting UAI uai_name = %s", uai_name)
         self.uas_cfg.get_config()
         if uai_name is None:
             abort(400, "Missing UAI Name argument")
-        candidate_list = self.select_deployments(
+        candidate_list = self.select_jobs(
             labels=["app=%s" % uai_name]
         )
         if not candidate_list:
@@ -107,26 +122,33 @@ class UasManager(UasBase):
                 404,
                 "no UAI information found for UAI '%s'" % uai_name
             )
+        logger.debug("got UAI: %s", resp_list[0])
         return resp_list[0]
 
     def get_uais(self, class_id=None, owner=None):
         """Get a list of UAIs optionally filtered on class and owner
 
         """
+        logger.debug(
+            "list UAIs class_id = %s, owner = %s",
+            class_id, owner
+        )
         self.uas_cfg.get_config()
         labels = []
         if owner is not None:
             labels.append("user=%s" % owner)
         if class_id is not None:
             labels.append("uas-class-id=%s" % class_id)
-        uai_list = self.select_deployments(labels=labels)
+        uai_list = self.select_jobs(labels=labels)
         resp_list = self.get_uai_list(uai_list)
+        logger.debug("found UAI list: %s", resp_list)
         return resp_list
 
     def delete_image(self, image_id):
         """Delete a UAI image from the config
 
         """
+        logger.debug("deleing UAI image '%s'", image_id)
         self.uas_cfg.get_config()
 
         # Make sure the image ID is not in use by any classes, and, if
@@ -148,16 +170,22 @@ class UasManager(UasBase):
         if img is None:
             abort(404, "image '%s' does not exist" % image_id)
         img.remove() # don't use img.delete() you actually want it removed
-        return {
+        ret = {
             'image_id': img.image_id,
             'imagename': img.imagename,
             'default': img.default
         }
+        logger.debug("deleted image: %s", ret)
+        return ret
 
     def create_image(self, imagename, default):
         """Create a new UAI image in the config
 
         """
+        logger.debug(
+            "creating (registering) UAI image '%s', default = %s",
+            imagename, default
+        )
         self.uas_cfg.get_config()
         if UAIImage.get_by_name(imagename):
             abort(409, "image named '%s' already exists" % imagename)
@@ -176,17 +204,23 @@ class UasManager(UasBase):
         # Now create the new image...
         img = UAIImage(imagename=imagename, default=default)
         img.put()
-        return {
+        ret = {
             'image_id': img.image_id,
             'imagename': img.imagename,
             'default': img.default
         }
+        logger.debug("created (registered) UAI image: %s", ret)
+        return ret
 
 
     def update_image(self, image_id, imagename, default):
         """Update a UAI image in the config
 
         """
+        logger.debug(
+            "updating image '%s' imagename = %s, default = %s",
+            image_id, imagename, default
+        )
         self.uas_cfg.get_config()
         img = UAIImage.get(image_id)
         if img is None:
@@ -224,35 +258,41 @@ class UasManager(UasBase):
                         tmp.default = False
                         tmp.put()
             img.put()
-        return {
+        ret = {
             'image_id': img.image_id,
             'imagename': img.imagename,
             'default': img.default
         }
+        logger.debug("Updated image %s: %s", image_id, ret)
+        return ret
 
     def get_image(self, image_id):
         """Retrieve a UAI image from the config
 
         """
+        logger.debug("get UAI image '%s'", image_id)
         self.uas_cfg.get_config()
         img = UAIImage.get(image_id)
         if img is None:
             abort(404, "image '%s' does not exist" % image_id)
-        return {
+        ret = {
             'image_id': img.image_id,
             'imagename': img.imagename,
             'default': img.default
         }
+        logger.debug("get UAI image '%s': %s", image_id, ret)
+        return ret
 
     def get_images(self):
         """Get the list of UAI images in the config
 
         """
+        logger.debug("list UAI images")
         self.uas_cfg.get_config()
         imgs = UAIImage.get_all()
         imgs = [] if imgs is None else imgs
         # pylint: disable=no-member
-        return [
+        ret = [
             {
                 'image_id': img.image_id,
                 'imagename': img.imagename,
@@ -260,11 +300,14 @@ class UasManager(UasBase):
             }
             for img in imgs
         ]
+        logger.debug("returning UAI image list: %s", ret)
+        return ret
 
     def delete_volume(self, volume_id):
         """Delete a UAI volume from the config
 
         """
+        logger.debug("deleting volume '%s'", volume_id)
         self.uas_cfg.get_config()
         # Make sure the volume ID is not in use by any classes, and, if
         # it is, get a list of them to complain about.
@@ -285,17 +328,23 @@ class UasManager(UasBase):
         if vol is None:
             abort(404, "volume '%s' does not exist" % volume_id)
         vol.remove() # don't use vol.delete() you actually want it removed
-        return {
+        ret = {
             'volume_id': vol.volume_id,
             'volumename': vol.volumename,
             'mount_path': vol.mount_path,
             'volume_description': vol.volume_description
         }
+        logger.debug("deleted volume '%s': %s", volume_id, ret)
+        return ret
 
     def create_volume(self, volumename, mount_path, vol_desc):
         """Create a UAI volume in the config
 
         """
+        logger.debug(
+            "creating volume '%s', mount_path = %s, vol_desc = %s",
+            volumename, mount_path, vol_desc
+        )
         self.uas_cfg.get_config()
         if not UAIVolume.is_valid_volume_name(volumename):
             abort(
@@ -332,18 +381,25 @@ class UasManager(UasBase):
             volume_description=vol_desc
         )
         vol.put()
-        return {
+        ret = {
             'volume_id': vol.volume_id,
             'volumename': vol.volumename,
             'mount_path': vol.mount_path,
             'volume_description': vol.volume_description
         }
+        logger.debug("created volume '%s': %s", volumename, ret)
+        return ret
 
     def update_volume(self, volume_id,
                       volumename=None, mount_path=None, vol_desc=None):
         """Update a UAI volume in the config
 
         """
+        logger.debug(
+            "updating volume '%s', volumename = %s,  mount_path = %s, "
+            "vol_desc = %s",
+            volume_id, volumename, mount_path, vol_desc
+        )
         self.uas_cfg.get_config()
         vol = UAIVolume.get(volume_id)
         if vol is None:
@@ -392,17 +448,20 @@ class UasManager(UasBase):
             changed = True
         if changed:
             vol.put()
-        return {
+        ret = {
             'volume_id': vol.volume_id,
             'volumename': vol.volumename,
             'mount_path': vol.mount_path,
             'volume_description': vol.volume_description
         }
+        logger.debug("updated volume '%s': %s", volume_id, ret)
+        return ret
 
     def get_volume(self, volume_id):
         """Get info on a specific volume from the config
 
         """
+        logger.debug("getting volume '%s'", volume_id)
         self.uas_cfg.get_config()
         vol = UAIVolume.get(volume_id)
         if vol is None:
@@ -410,22 +469,25 @@ class UasManager(UasBase):
                 404,
                 "Unknown volume '%s'" % volume_id
             )
-        return {
+        ret = {
             'volume_id': vol.volume_id,
             'volumename': vol.volumename,
             'mount_path': vol.mount_path,
             'volume_description': vol.volume_description
         }
+        logger.debug("got volume '%s': %s", volume_id, ret)
+        return ret
 
     def get_volumes(self):
         """Get info on all volumes in the config
 
         """
+        logger.debug("listing volumes")
         self.uas_cfg.get_config()
         vols = UAIVolume.get_all()
         vols = [] if vols is None else vols
         # pylint: disable=no-member
-        return [
+        ret = [
             {
                 'volume_id': vol.volume_id,
                 'volumename': vol.volumename,
@@ -434,11 +496,14 @@ class UasManager(UasBase):
             }
             for vol in vols
         ]
+        logger.debug("found the following volumes: %s", ret)
+        return ret
 
     def delete_resource(self, resource_id):
         """Delete resource limit / request config
 
         """
+        logger.debug("deleting resource '%s'", resource_id)
         self.uas_cfg.get_config()
 
         # Make sure the resource is not in use by any classes, and, if
@@ -464,17 +529,23 @@ class UasManager(UasBase):
 
         # Good to go...
         resource.remove() # don't use x.delete() you actually want it removed
-        return {
+        ret = {
             'resource_id': resource.resource_id,
             'comment': resource.comment,
             'limit': resource.limit,
             'request': resource.request
         }
+        logger.debug("deleted resource '%s': %s", resource_id, ret)
+        return ret
 
     def create_resource(self, comment=None, limit=None, request=None):
         """Create a UAI resource limit / request config
 
         """
+        logger.debug(
+            "creating resource, comment=%s, limit=%s, request = %s",
+            comment, limit, request
+        )
         self.uas_cfg.get_config()
         # If 'limit' is specified convert it from a JSON string to a dictionary
         if limit is not None:
@@ -504,12 +575,14 @@ class UasManager(UasBase):
             request=request
         )
         resource.put()
-        return {
+        ret = {
             'resource_id': resource.resource_id,
             'comment': resource.comment,
             'limit': resource.limit,
             'request': resource.request
         }
+        logger.debug("created resource: %s", ret)
+        return ret
 
     def update_resource(self, resource_id,
                         comment=None,
@@ -518,6 +591,10 @@ class UasManager(UasBase):
         """Update a resource limit / request config
 
         """
+        logger.debug(
+            "updating resource '%s' comment=%s, limit=%s, request = %s",
+            resource_id, comment, limit, request
+        )
         self.uas_cfg.get_config()
         resource = UAIResource.get(resource_id)
         if resource is None:
@@ -553,17 +630,20 @@ class UasManager(UasBase):
             changed = True
         if changed:
             resource.put()
-        return {
+        ret = {
             'resource_id': resource.resource_id,
             'comment': resource.comment,
             'limit': resource.limit,
             'request': resource.request
         }
+        logger.debug("updated resource '%s': %s", resource_id, ret)
+        return ret
 
     def get_resource(self, resource_id):
         """Get info on a specific resource limit / request config
 
         """
+        logger.debug("getting resource '%s'", resource_id)
         self.uas_cfg.get_config()
         resource = UAIResource.get(resource_id)
         if resource is None:
@@ -571,22 +651,25 @@ class UasManager(UasBase):
                 404,
                 "Unknown resource '%s'" % resource_id
             )
-        return {
+        ret = {
             'resource_id': resource.resource_id,
             'comment': resource.comment,
             'limit': resource.limit,
             'request': resource.request
         }
+        logger.debug("got resource '%s': %s", resource_id, ret)
+        return ret
 
     def get_resources(self):
         """Get info on all resource limit / request configs
 
         """
+        logger.debug("listing resources")
         self.uas_cfg.get_config()
         resources = UAIResource.get_all()
         resources = [] if resources is None else resources
         # pylint: disable=no-member
-        return [
+        ret = [
             {
                 'comment': resource.comment,
                 'limit': resource.limit,
@@ -594,6 +677,8 @@ class UasManager(UasBase):
             }
             for resource in resources
         ]
+        logger.debug("got list of resources: %s", ret)
+        return ret
 
     @staticmethod
     def _validate_volume_list(volume_list):
@@ -656,6 +741,74 @@ class UasManager(UasBase):
                     )
 
     @staticmethod
+    def _validate_timeout(timeout):
+        """Verify that a given timeout description is a validly
+        formed dictionary of timeout settings.
+
+        """
+        if timeout is not None:
+            try:
+                timeout_dict = json.loads(timeout)
+            except json.decoder.JSONDecodeError as err:
+                abort(
+                    400,
+                    "Timeout '%s' failed JSON decoding "
+                    "- %s" % (timeout, str(err))
+                )
+            if not isinstance(timeout_dict, dict):
+                abort(
+                    400,
+                    "Timeout '%s' must be a JSON map object "
+                    "but is not a map" %
+                    (timeout)
+                )
+            for key, value in timeout_dict.items():
+                if key not in ['soft', 'hard', 'warning']:
+                    abort(
+                        400,
+                        "Timeout '%s' contains an unrecognized timeout "
+                        "setting '%s' acceptable settings are "
+                        "'soft', 'hard' or 'warning'" % (timeout, key)
+                    )
+                if not isinstance(value, str):
+                    abort(
+                        400,
+                        "Timeout '%s' setting '%s: %s' has a non-string "
+                        "value" % (timeout, key, value)
+                    )
+                try:
+                    intval = int(value)
+                except ValueError as err:
+                    abort(
+                        400,
+                        "Timeout '%s' setting '%s: %s' cannot be "
+                        "converted to an integer - %s" %
+                        (timeout, key, value, err)
+                    )
+                if intval < 0:
+                    abort(
+                        400,
+                        "Timeout '%s' setting '%s: %s' must not be "
+                        "a negative value" %
+                        (timeout, key, value)
+                    )
+
+
+    @staticmethod
+    def _validate_service_account(service_account):
+        """Verify that a given service account name is a valid Kubernetes
+        name.
+
+        """
+        valid_re = re.compile(r"^[a-zA-Z0-9-]+$")
+        if valid_re.match(service_account) is not None:
+            abort(
+                400,
+                "Invalid service account name '%s'" %
+                service_account
+            )
+
+    @staticmethod
     def _expanded_uai_class(uai_class):
         """Fully expand a UAI Class object and all of its sub-objects.  This
         differs from the object based `expand` method used elsewhere
@@ -696,19 +849,24 @@ class UasManager(UasBase):
             'uai_image': UAIImage.get(uai_class.image_id).expand(),
             'resource_config': resource_config,
             'volume_mounts': volume_mounts,
-            'tolerations': uai_class.tolerations
+            'tolerations': uai_class.tolerations,
+            'timeout': uai_class.timeout,
+            'service_account': uai_class.service_account,
         }
 
     def delete_class(self, class_id):
         """Delete a UAI Class
 
         """
+        logger.debug("deleting UAI class '%s'", class_id)
         self.uas_cfg.get_config()
         uai_class = UAIClass.get(class_id)
         if uai_class is None:
             abort(404, "UAI Class '%s' does not exist" % class_id)
         uai_class.remove() # don't use x.delete() you actually want it removed
-        return self._expanded_uai_class(uai_class)
+        ret = self._expanded_uai_class(uai_class)
+        logger.debug("deleted UAI class '%s': %s", class_id, ret)
+        return ret
 
     #pylint: disable=too-many-arguments,too-many-statements,too-many-locals
     def create_class(self,
@@ -723,10 +881,24 @@ class UasManager(UasBase):
                      uai_compute_network=None,
                      resource_id=None,
                      volume_list=None,
-                     tolerations=None):
+                     tolerations=None,
+                     timeout=None,
+                     service_account=None):
         """Create a UAI Class
 
         """
+        logger.debug(
+            "creating UAI class, comment = %s, default = %s, public_ip = %s, "
+            "image_id = %s, priority_class_name = %s, namespace = %s, "
+            "opt_ports = %s, uai_creation_class = %s, "
+            "uai_compute_network = %s, resource_id = %s, volume_list = %s, "
+            "tolerations = %s, timeout = %s, "
+            "service_account = %s",
+            comment, default, public_ip, image_id, priority_class_name,
+            namespace, opt_ports, uai_creation_class, uai_compute_network,
+            resource_id, volume_list, tolerations, timeout,
+            service_account
+        )
         self.uas_cfg.get_config()
         if image_id is None:
             abort(400, "Must specify an image ID when creating a UAI Class")
@@ -749,6 +921,9 @@ class UasManager(UasBase):
             )
         if volume_list:
             self._validate_volume_list(volume_list)
+        if timeout:
+            self._validate_timeout(timeout)
+        timeout = json.loads(timeout) if timeout is not None else None
         opt_ports_list = [
             port.strip()
             for port in opt_ports.split(',')
@@ -768,10 +943,14 @@ class UasManager(UasBase):
         comment = "" if comment is None else comment
         default = False if default is None else default
         public_ip = False if public_ip is None else public_ip
-        uai_compute_network = True if uai_compute_network is None else uai_compute_network
-        priority_class_name = ("uai-priority"
-                               if priority_class_name is None
-                               else priority_class_name)
+        uai_compute_network = (
+            True if uai_compute_network is None
+            else uai_compute_network
+        )
+        priority_class_name = (
+            "uai-priority" if priority_class_name is None
+            else priority_class_name
+        )
         namespace = (
             self.uas_cfg.get_uai_namespace()
             if namespace is None
@@ -791,6 +970,8 @@ class UasManager(UasBase):
             resource_id=resource_id,
             volume_list=volume_list,
             tolerations=tolerations,
+            timeout=timeout,
+            service_account=service_account
         )
         if default:
             default_class = UAIClass.get_default()
@@ -801,7 +982,9 @@ class UasManager(UasBase):
                 default_class.default = False
                 default_class.put()  # pylint: disable=no-member
         uai_class.put()
-        return self._expanded_uai_class(uai_class)
+        ret = self._expanded_uai_class(uai_class)
+        logger.debug("created UAI class: %s", ret)
+        return ret
 
     # pylint: disable=too-many-branches,too-many-locals
     def update_class(self,
@@ -817,10 +1000,24 @@ class UasManager(UasBase):
                      uai_compute_network=None,
                      resource_id=None,
                      volume_list=None,
-                     tolerations=None):
+                     tolerations=None,
+                     timeout=None,
+                     service_account=None):
         """Update a UAI Class
 
         """
+        logger.debug(
+            "updating UAI class '%s', comment = %s, default = %s, "
+            "public_ip = %s, image_id = %s, priority_class_name = %s, "
+            "namespace = %s, opt_ports = %s, uai_creation_class = %s, "
+            "uai_compute_network = %s, resource_id = %s, volume_list = %s, "
+            "tolerations = %s, timeout = %s, "
+            "service_account = %s",
+            class_id, comment, default, public_ip, image_id,
+            priority_class_name, namespace, opt_ports, uai_creation_class,
+            uai_compute_network, resource_id, volume_list, tolerations,
+            timeout, service_account
+        )
         self.uas_cfg.get_config()
         uai_class = UAIClass.get(class_id)
         if uai_class is None:
@@ -898,6 +1095,14 @@ class UasManager(UasBase):
             self._validate_tolerations(tolerations)
             uai_class.tolerations = tolerations
             changed = True
+        if timeout is not None:
+            self._validate_timeout(timeout)
+            uai_class.timeout = json.loads(timeout)
+            changed = True
+        if service_account is not None:
+            self._validate_service_account(service_account)
+            uai_class.service_account = service_account
+            changed = True
         if changed:
             if default:  # this implies that default is not None
                 default_class = UAIClass.get_default()
@@ -908,12 +1113,15 @@ class UasManager(UasBase):
                     default_class.default = False
                     default_class.put()
             uai_class.put()
-        return self._expanded_uai_class(uai_class)
+        ret =  self._expanded_uai_class(uai_class)
+        logger.debug("updated UAI class '%s': %s", class_id, ret)
+        return ret
 
     def get_class(self, class_id):
         """Get info on a specific class limit / request config
 
         """
+        logger.debug("getting UAI class '%s'", class_id)
         self.uas_cfg.get_config()
         uai_class = UAIClass.get(class_id)
         if uai_class is None:
@@ -921,25 +1129,31 @@ class UasManager(UasBase):
                 404,
                 "Unknown class '%s'" % class_id
             )
-        return self._expanded_uai_class(uai_class)
+        ret = self._expanded_uai_class(uai_class)
+        logger.debug("got UAI class '%s': %s", class_id, ret)
+        return ret
 
     def get_classes(self):
         """Get info on all class limit / request configs
 
         """
+        logger.debug("listing UAI classes")
         self.uas_cfg.get_config()
         uai_classes = UAIClass.get_all()
         uai_classes = [] if uai_classes is None else uai_classes
-        return [
+        ret = [
             self._expanded_uai_class(uai_class)
             for uai_class in uai_classes
         ]
+        logger.debug("got list of UAI classes: %s", ret)
+        return ret
 
     def factory_reset(self):
         """Delete all the local configuration so that the next operation
         reloads config from the configmap configuration.
 
         """
+        logger.debug("resetting UAS config to factory defaults")
         self.uas_cfg.get_config()
         vols = UAIVolume.get_all()
         vols = [] if vols is None else vols
@@ -961,3 +1175,4 @@ class UasManager(UasBase):
         cfgs = [] if cfgs is None else cfgs
         for cfg in cfgs:
             cfg.remove()
+        logger.debug("UAS config has been reset to factory defaults")